+++ NEWS.md
@@ -4,6 +4,8 @@
 
 * new `tm_t_summary_by` module
 
+* add stratified analysis to `tm_g_forest_tte` and `tm_g_forest_rsp`
+
 # teal.modules.clinical 0.8.1
 
 * new `tm_t_events` and `tm_t_events_by_grade` modules